# Performance improvements to CoberturaParser